Move user list functions in another class to be able to test
